# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Platform Messaging Adapter API
  description: |
    RESTful API for sending and managing messages across multiple platforms:
    - **WhatsApp** (via Infobip, Twilio)
    - **Instagram** (via Meta Messenger Platform API)
    - **Facebook Messenger** (via Meta Messenger Platform API)

    This API provides unified endpoints for messaging, template management, webhooks, and monitoring.
  version: 1.0.0
  contact:
    name: API Support
    email: support@your-domain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your-domain.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /api/messages/text:
    post:
      tags:
        - Messages
      summary: Send text message
      description: Send a free-form text message (WhatsApp, Instagram, Messenger)
      operationId: sendText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextRequest"
      responses:
        "200":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "429":
          $ref: "#/components/responses/RateLimitError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/messages/media:
    post:
      tags:
        - Messages
      summary: Send media message
      description: Send media (image, document, audio, or video)
      operationId: sendMedia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaRequest"
      responses:
        "200":
          description: Media sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/RateLimitError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/messages/interactive/buttons:
    post:
      tags:
        - Messages
      summary: Send interactive buttons
      description: Send an interactive message with buttons (max 3 buttons for WhatsApp, max 13 for Meta)
      operationId: sendInteractiveButtons
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractiveButtonsRequest"
      responses:
        "200":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/RateLimitError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key authentication using Bearer token
...
  responses:
    BadRequestError:
      description: Bad request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid phone number format
            timestamp: "2026-01-20T10:30:00Z"
    UnauthorizedError:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid or missing API key
            timestamp: "2026-01-20T10:30:00Z"
    ForbiddenError:
      description: Forbidden - Session expired or permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: SESSION_EXPIRED
              message: WhatsApp session expired (use HSM to start new session)
            timestamp: "2026-01-20T10:30:00Z"
...
    RateLimitError:
      description: Too many requests - Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests
            timestamp: "2026-01-20T10:30:00Z"
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests allowed
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the limit resets
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: Internal server error
            timestamp: "2026-01-20T10:30:00Z"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid parameters
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2026-01-20T10:30:00Z"

    # Health Schemas
...
    TextRequest:
      type: object
      required:
        - to
        - text
      properties:
        to:
          type: string
          description: Phone number (E.164) or IGSID/PSID for Meta
          example: "+351912345678"
        text:
          type: string
          description: Message text (max 2000 characters)
          maxLength: 2000
          example: Hello! How can I help you today?
        previewUrl:
          type: boolean
          description: Enable URL preview (WhatsApp only)
          default: false
        provider:
          type: string
          enum: [infobip, twilio, meta]
          example: infobip
        platform:
          type: string
          enum: [instagram, messenger]
          description: Required for Meta provider
          example: instagram
    MediaRequest:
      type: object
      required:
        - to
        - type
        - url
      properties:
        to:
          type: string
          description: Phone number (E.164) or IGSID/PSID for Meta
          example: "+351912345678"
        type:
          type: string
          enum: [image, document, audio, video]
          example: image
        url:
          type: string
          format: uri
          description: Public URL of the media file
          example: https://example.com/image.jpg
        caption:
          type: string
          description: Media caption (optional)
          maxLength: 1024
          example: Check out this image!
        filename:
          type: string
          description: Filename for documents
          example: document.pdf
        provider:
          type: string
          enum: [infobip, twilio, meta]
          example: infobip
        platform:
          type: string
          enum: [instagram, messenger]
          description: Required for Meta provider
    InteractiveButtonsRequest:
      type: object
      required:
        - to
        - body
        - buttons
      properties:
        to:
          type: string
          example: "+351912345678"
        body:
          type: string
          description: Message body text
          example: Please select an option
        buttons:
          type: array
          description: Buttons (max 3 for WhatsApp, max 13 for Meta)
          minItems: 1
          maxItems: 13
          items:
            type: object
            required:
              - id
              - text
            properties:
              id:
                type: string
                example: btn_1
              text:
                type: string
                maxLength: 20
                example: Option 1
        headerText:
          type: string
          description: Optional header text
        footerText:
          type: string
          description: Optional footer text
        provider:
          type: string
          enum: [infobip, twilio, meta]
        platform:
          type: string
          enum: [instagram, messenger]
...
    MessageResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - messageId
            - status
          properties:
            messageId:
              type: string
              example: msg_abc123
            status:
              type: string
              enum: [PENDING, SENT, DELIVERED, READ, FAILED]
              example: PENDING
            to:
              type: string
              example: "+351912345678"
            sentAt:
              type: string
              format: date-time
              example: "2026-01-20T10:30:00Z"
            platform:
              type: string
              enum: [whatsapp, instagram, messenger]
            messagingWindowExpiresAt:
              type: string
              format: date-time
              description: For Meta platforms only
        timestamp:
          type: string
          format: date-time
...